Update note on bad regression (31.3.0 9-3-2022) (#57)
Hi Breck (@breck7),

A week ago you were under a lot of pressure and were castigating yourself on a problem with the code. These may be thoughts that you hold privately, since you hold yourself to a high standard of coding. But we all make mistakes, and you are doing a mountain of work, of which you should be proud, even if mistakes are inevitable.

Some of your pull requests also have interesting descriptions! 😄 However, the Release Notes document is not a place for disparaging your own work. 

My 2¢.

Kind Regards,
Liam